Add initial version of translation program
This change adds the initial version of the dict-cc-cli program. The
program provides a very simple command line interface for the
translation from English to German (that really depends on the database
being used but these are the intended languages).
The program relies on a sqlite database containing the necessary
translations. Such a database can be copied, for example, from the
Android version of the dict.cc app.
Right now the program is only capable of translating exact matches,
i.e., words that are contained in the database in precisely that form.