Add tests for translation functionality
This change adds tests for the translation infrastructure. In order
to allow for proper testing the `translate` function got reworked
slightly to accept the database to use as well as a callback function
that is being invoked for each translation.
Also, for us to work with a sqlite database in the correct format we
use a slightly modified version of the original (small) one. In a
nutshell we remove unused data to make the file nicely compressible.