Correctly handle non-existent database
The sqlite crate is not to be used only for querying databases but also
for modifying them. As such, when an attempt is made to open a database
that does not yet exist it will be created.
This behavior is not desired in our case as we always need an existing
database and would like to avoid the creation of empty ones. To that
end, this change checks if the database file exists before opening it
and returns an appropriate error should that not be the case.