Display English term being translated as well
So far we performed a match for equality when looking for terms to
translate and their matches. That approach is insufficient for various
reasons and will be replaced with a more elaborate scheme that cleverly
checks certain substrings in the future.
In order to be prepared for such a future change, this change adjusts
the code slightly to also pass the actual term being translated to the
callback function, which, in the non-test case will print it along the
translation.