Handle inconsistent spacing for meta data suffixes
As it turned out, some of the words that contain a meta data suffix have
inconsistent spacing between the term and the suffix. "dorky", for
example, exists in the database as "dorky [coll.]" and "dorky  [coll.]".
In order to deal with those cases we add an additional OR condition to
our SQL query.
Unfortunately, that is not a great solution as other or different white
spaces would still cause insufficient results to be reported, but the
limits imposed on us by sqlite (namely, no regular expression support in
conjunction with a limited pattern matching language not allowing to
match for a variable number of occurrences of the same character) make
it seemingly impossible to have a holistic solution to this problem.