Handle undefined types of terms properly
Some terms to translate do not have an associated type. "love", for
example, as used in Tennis, is simply untyped in the database. To make
these cases a tad more explicit (and fix the weird looking output of the
program for those cases) this change replaces the empty string reported
as part of the SQL query with "unknown".