Mitigate SQL injection problems
So far we have taken the user input quite literally and embedded it
directly into the SQL statement to execute. That is a problem because
maliciously crafted input can lead to SQL injection attacks. Although
arguably not a big issue in this program because the user can safely be
assumed to not have malicious intent, the code should be written in a
way that prevents such attacks in the first place.
This change mitigates the problem by *correctly* using a prepared
statement. Instead of using our user input and inserting it into the SQL
query using a string format function, we now use sqlite's bind
functionality to safely embed the user provided string into the query.