Always search in a case-insensitive manner
The equality comparison provided by SQL is case sensitive whereas the
LIKE operator is not. Because we used an equality check to find exact
matches we were unable to find certain ones if they started with a
capital letter.
This change adjusts the program to use the LIKE operator in all
selections.