Handle multi word translations correctly
So far we have focused solely on single word translations (although in
some variations to accommodate the format used in the database). With
this change we finally support multi word translations, e.g., those
where only one word of a term matches the input.
Through use of a UNION construct we are able to seemingly keep a stable
sort order with results from the first query appearing before results
from the second one. Note that early versions of this change used a LEFT
JOIN to create the final set. However, it was found that this approach
way slower than the current one.