Require database to use to be passed in
In its current state the program is not usable in a flexible manner due
to the hardcoded path data/dictcc-lp1.db of the database relative to the
binary:
- requiring the database to reside in a data/ subdirectory is a no-go
  when it comes to a global installation
- it is also not (easily) possible to use different databases

This change requires the database to use to be passed in as the first
argument. It also allows for a more intuitive handling of additional
arguments, which will now be treated as the words to search for (as
opposed to requiring the user to cram the exact search string into the
very first argument).