Use getopts crate for argument parsing
So far we have rolled our own argument parsing. That was fine because we
only had positional arguments. With the upcoming introduction of
additional functionality we need support for a flag parameters. Because
POSIX compliant handling of program flags is tricky to get right, this
change pulls in and uses the getopts crate for the task of parsing
arguments.
The clap crate was evaluated and despite having the arguably nicer
interface its higher number of dependencies and the extreme code bloat
it causes (from 345 KiB to 888 KiB; even in version 2.29.1, which
supposedly contains already fixes to address bloat) are a no-go.