Introduce support for bidirectional translations
This change introduces support for bidirectional translations (which, as
it stands is still not hooked up to the command line interface). In
particular, it introduces an enum describing the direction in which we
translate (either from language 1 to language 2 or vice versa). It also
adjusts the `translate` function to dynamically select the source and
destination columns to work with based on such a value.