Add test for case-insensitive matching of Unicode characters
During usage a bug was discovered which manifested itself as certain
Unicode characters, German Umlaute in particular, to be not matched in a
case-insensitive manner. The noun Ärger, for example, would only match
translations that are stored as beginning with a capital Ä. The same was
found to be true for words beginning with a lower cased letter.
Investigation has shown that this issue is not with the program itself,
but rather with sqlite. More specifically, if the system's sqlite
library is not built with ICU (Internationalization Components for
Unicode) support, this behavior is exhibited.
As dictcc-cli is meant to be a program to be used with various
languages, with a lot of them using characters beyond the ASCII range,
we consider this behavior generally undesirable. To that end, this
change adds a test that triggers this case. The test will fail if the
system's sqlite library is not built with support for ICU to provide an
early indication of this problem.